# docs/swagger/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Server API",
    "version": "1.0.0",
    "description": "These are the endpoints for the Connector Server API",
    "contact": {
      "name": "Alessandro Cannarella",
      "email": "cannarella.dev@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "basePath": "/api",
  "schemes": [
    "http"
  ],
...
  "paths": {
...
    "/peer": {
      "post": {
        "tags": [
          "peer"
        ],
        "summary": "Create a peering",
        "operationId": "createPeering",
        "description": "Create a peering with a remote cluster based on an existing contract",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "description": "Cluster ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "description": "Cluster name",
            "required": true,
            "type": "string"
          },
          {
            "name": "authtoken",
            "in": "query",
            "description": "Cluster token",
            "required": true,
            "type": "string"
          },
          {
            "name": "authurl",
            "in": "query",
            "description": "Cluster auth url",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "get": {
        "deprecated": true,
        "tags": [
          "peer"
        ],
        "summary": "Get peerings",
        "description": "Get existing peerings - [NOT IMPLEMENTED]",
        "operationId": "getPeerings",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },
      "delete": {
        "deprecated": true,
        "tags": [
          "peer"
        ],
        "summary": "Delete a peering",
        "description": "Delete a peering with a remote cluster - [NOT IMPLEMENTED]",
        "operationId": "deletePeering",
        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/contracts/buy": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Buy a contract",
        "description": "Buy an available offer plan from a foreign cluster and stipulate a contract with it",
        "operationId": "buyContract",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "plan-id",
            "in": "query",
            "description": "Plan ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "offer-id",
            "in": "query",
            "description": "Offer ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ContractDocument"
                }
              }
            }
          }
        }
      }
    },
    "/contracts/sell": {
      "post": {
        "tags": [
          "contracts"
        ],
        "summary": "Sell a contract",
        "description": "Sell an available offer plan to a foreign cluster and stipulate a contract with it",
        "operationId": "sellContract",
        "consumes": [
          "multipart/form-data"
        ],
        "parameters": [
          {
            "name": "plan-id",
            "in": "query",
            "description": "Plan ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "offer-id",
            "in": "query",
            "description": "Offer ID",
            "required": true,
            "type": "string"
          },
          {
            "name": "buyer-id",
            "in": "query",
            "description": "Buyer ID",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/definitions/ContractDocument"
                }
              }
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "Offer": {
      "type": "object",
      "properties": {
        "offerID": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426655440000"
        },
        "offerName": {
          "type": "string",
          "example": "Example Offer"
        },
        "offerType": {
          "type": "string",
          "example": "computational"
        },
        "description": {
          "type": "string",
          "example": "Example Offer Description"
        },
        "plans": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Plan"
          }
        },
        "clusterPrettyName": {
          "type": "string",
          "example": "Example Cluster"
        },
        "created": {
          "type": "integer",
          "example": "1234567890"
        },
        "status": {
          "type": "boolean",
          "example": true
        }
      }
    },
    "Plan": {
      "type": "object",
      "properties": {
        "planID": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426655440000_1"
        },
        "planName": {
          "type": "string",
          "example": "Basic"
        },
        "planCost": {
          "type": "number",
          "example": 50.99
        },
        "planCostCurrency": {
          "type": "string",
          "example": "USD"
        },
        "planCostPeriod": {
          "type": "string",
          "example": "month"
        },
        "planQuantity": {
          "type": "integer",
          "example": 10
        },
        "resources": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      }
    },
    "ContractDocument": {
      "type": "object",
      "properties": {
        "contractID": {
          "type": "string",
          "example": "5d6a3b9e-9f9e-4a0e-8d1a-5a0f0a2e1a0a"
        },
        "buyerID": {
          "type": "string",
          "example": "5d6a3b9e-9f9e-4a0e-8d1a-5a0f0a2e1a0b"
        },
        "seller": {
          "$ref": "#/definitions/Provider"
        },
        "offer": {
          "$ref": "#/definitions/Offer"
        },
        "planID": {
          "type": "string",
          "example": "123e4567-e89b-12d3-a456-426655440000_1"
        },
        "enabled": {
          "type": "boolean",
          "example": true
        },
        "created": {
          "type": "integer",
          "example": "1234567890"
        }
      },
      "required": [
        "contractID",
        "buyerID",
        "seller",
        "offer",
        "planID",
        "enabled",
        "created"
      ]
    },
    "Provider": {
      "type": "object",
      "properties": {
        "clusterPrettyName": {
          "type": "string",
          "example": "Example Cluster"
        },
        "clusterContractEndpoint": {
          "type": "string",
          "example": "http://example.com/contracts"
        },
        "clusterName": {
          "type": "string",
          "example": "example-cluster"
        },
        "clusterID": {
          "type": "string",
          "example": "7d6a3b9e-9f9e-4a0e-8d1a-5a0f0a2e1a0a"
        },
        "endpoint": {
          "type": "string",
          "example": "http://example.com:8080"
        },
        "token": {
          "type": "string",
          "example": "1234567890123456789012345678901234567890123456789012345678901234"
        }
      },
      "required": [
        "clusterPrettyName",
        "clusterContractEndpoint",
        "ClusterParameters"
      ]
    },
...
  }
}